Change tests

+++ tests/auth_test.py
@@ -8,12 +8,10 @@ import unittest
 import signal
 import time
 
-from .helpers.ptrack_helpers import ProbackupTest, ProbackupException
+from .helpers.ptrack_helpers import ProbackupTest
 from testgres import StartNodeException
 
-module_name = 'auth_test'
 skip_test = False
-
 
 try:
     from pexpect import *
@@ -21,7 +19,7 @@ except ImportError:
     skip_test = True
 
 
-class SimpleAuthTest(ProbackupTest, unittest.TestCase):
+class SimpleAuthTest(ProbackupTest):
 
     # @unittest.skip("skip")
     def test_backup_via_unprivileged_user(self):
@@ -30,16 +28,13 @@ class SimpleAuthTest(ProbackupTest, unittest.TestCase):
             run a backups without EXECUTE rights on
             certain functions
         """
-        node = self.make_simple_node(
-            base_dir=os.path.join(self.module_name, self.fname, 'node'),
+        node = self.pg_node.make_simple('node',
             set_replication=True,
-            ptrack_enable=self.ptrack,
-            initdb_params=['--data-checksums'])
+            ptrack_enable=self.ptrack)
 
-        backup_dir = os.path.join(self.tmp_path, self.module_name, self.fname, 'backup')
-        self.init_pb(backup_dir)
-        self.add_instance(backup_dir, 'node', node)
-        self.set_archiving(backup_dir, 'node', node)
+        self.pb.init()
+        self.pb.add_instance('node', node)
+        self.pb.set_archiving('node', node)
         node.slow_start()
 
         if self.ptrack:
@@ -49,27 +44,18 @@ class SimpleAuthTest(ProbackupTest, unittest.TestCase):
 
         node.safe_psql("postgres", "CREATE ROLE backup with LOGIN")
 
-        try:
-            self.backup_node(
-                backup_dir, 'node', node, options=['-U', 'backup'])
-            self.assertEqual(
-                1, 0,
-                "Expecting Error due to missing grant on EXECUTE.")
-        except ProbackupException as e:
-            if self.get_version(node) < 150000:
-                self.assertIn(
-                    "ERROR: query failed: ERROR:  permission denied "
-                    "for function pg_start_backup", e.message,
-                    '\n Unexpected Error Message: {0}\n CMD: {1}'.format(
-                        repr(e.message), self.cmd))
-            else:
-                self.assertIn(
-                    "ERROR: query failed: ERROR:  permission denied "
-                    "for function pg_backup_start", e.message,
-                    '\n Unexpected Error Message: {0}\n CMD: {1}'.format(
-                        repr(e.message), self.cmd))
+        self.pb.backup_node('node', node, options=['-U', 'backup'],
+            expect_error='due to missing grant on EXECUTE')
+        if self.pg_config_version < 150000:
+            self.assertMessage(contains=
+                "ERROR: Query failed: ERROR:  permission denied "
+                "for function pg_start_backup")
+        else:
+            self.assertMessage(contains=
+                "ERROR: Query failed: ERROR:  permission denied "
+                "for function pg_backup_start")
 
-        if self.get_version(node) < 150000:
+        if self.pg_config_version < 150000:
             node.safe_psql(
                 "postgres",
                 "GRANT EXECUTE ON FUNCTION"
@@ -80,59 +66,36 @@ class SimpleAuthTest(ProbackupTest, unittest.TestCase):
                 "GRANT EXECUTE ON FUNCTION"
                 " pg_backup_start(text, boolean) TO backup;")
 
-        if self.get_version(node) < 100000:
-            node.safe_psql(
-                'postgres',
-                "GRANT EXECUTE ON FUNCTION pg_catalog.pg_switch_xlog() TO backup")
-        else:
-            node.safe_psql(
-                'postgres',
-                "GRANT EXECUTE ON FUNCTION pg_catalog.pg_switch_wal() TO backup")
+        node.safe_psql(
+            'postgres',
+            "GRANT EXECUTE ON FUNCTION pg_catalog.pg_switch_wal() TO backup")
 
-        try:
-            self.backup_node(
-                backup_dir, 'node', node, options=['-U', 'backup'])
-            self.assertEqual(
-                1, 0,
-                "Expecting Error due to missing grant on EXECUTE.")
-        except ProbackupException as e:
-            self.assertIn(
-                "ERROR: query failed: ERROR:  permission denied for function "
-                "pg_create_restore_point\nquery was: "
-                "SELECT pg_catalog.pg_create_restore_point($1)", e.message,
-                '\n Unexpected Error Message: {0}\n CMD: {1}'.format(
-                    repr(e.message), self.cmd))
+        self.pb.backup_node('node', node,
+                         options=['-U', 'backup'],
+                         expect_error='due to missing grant on EXECUTE')
+        self.assertMessage(contains=
+            "ERROR: Query failed: ERROR:  permission denied for function "
+            "pg_create_restore_point\nquery was: "
+            "SELECT pg_catalog.pg_create_restore_point($1)")
 
         node.safe_psql(
             "postgres",
             "GRANT EXECUTE ON FUNCTION"
             " pg_create_restore_point(text) TO backup;")
 
-        try:
-            self.backup_node(
-                backup_dir, 'node', node, options=['-U', 'backup'])
-            self.assertEqual(
-                1, 0,
-                "Expecting Error due to missing grant on EXECUTE.")
-        except ProbackupException as e:
-            if self.get_version(node) < 150000:
-                self.assertIn(
-                    "ERROR: Query failed: ERROR:  permission denied "
-                    "for function pg_stop_backup", e.message,
-                    '\n Unexpected Error Message: {0}\n CMD: {1}'.format(
-                        repr(e.message), self.cmd))
-            else:
-                self.assertIn(
-                    "ERROR: Query failed: ERROR:  permission denied "
-                    "for function pg_backup_stop", e.message,
-                    '\n Unexpected Error Message: {0}\n CMD: {1}'.format(
-                        repr(e.message), self.cmd))
+        self.pb.backup_node('node', node,
+                         options=['-U', 'backup'],
+                         expect_error='due to missing grant on EXECUTE')
+        if self.pg_config_version < 150000:
+            self.assertMessage(contains=
+                "ERROR: Query failed: ERROR:  permission denied "
+                "for function pg_stop_backup")
+        else:
+            self.assertMessage(contains=
+                "ERROR: Query failed: ERROR:  permission denied "
+                "for function pg_backup_stop")
 
-        if self.get_version(node) < self.version_to_num('10.0'):
-            node.safe_psql(
-                "postgres",
-                "GRANT EXECUTE ON FUNCTION pg_stop_backup(boolean) TO backup")
-        elif self.get_version(node) < self.version_to_num('15.0'):
+        if self.pg_config_version < self.version_to_num('15.0'):
             node.safe_psql(
                 "postgres",
                 "GRANT EXECUTE ON FUNCTION pg_stop_backup() TO backup; "
@@ -142,13 +105,11 @@ class SimpleAuthTest(ProbackupTest, unittest.TestCase):
                 "postgres",
                 "GRANT EXECUTE ON FUNCTION pg_backup_stop(boolean) TO backup;")
 
-        self.backup_node(
-                backup_dir, 'node', node, options=['-U', 'backup'])
+        self.pb.backup_node('node', node, options=['-U', 'backup'])
 
         node.safe_psql("postgres", "CREATE DATABASE test1")
 
-        self.backup_node(
-                backup_dir, 'node', node, options=['-U', 'backup'])
+        self.pb.backup_node('node', node, options=['-U', 'backup'])
 
         node.safe_psql(
             "test1", "create table t1 as select generate_series(0,100)")
@@ -161,64 +122,44 @@ class SimpleAuthTest(ProbackupTest, unittest.TestCase):
                 "ALTER ROLE backup REPLICATION")
 
         # FULL
-        self.backup_node(
-            backup_dir, 'node', node, options=['-U', 'backup'])
+        self.pb.backup_node('node', node, options=['-U', 'backup'])
 
         # PTRACK
         if self.ptrack:
-            self.backup_node(
-                backup_dir, 'node', node,
+            self.pb.backup_node('node', node,
                 backup_type='ptrack', options=['-U', 'backup'])
 
 
-class AuthTest(unittest.TestCase):
+class AuthTest(ProbackupTest):
     pb = None
     node = None
 
     # TODO move to object scope, replace module_name
-    @classmethod
-    def setUpClass(cls):
+    @unittest.skipIf(skip_test, "Module pexpect isn't installed. You need to install it.")
+    def setUp(self):
 
-        super(AuthTest, cls).setUpClass()
+        super().setUp()
 
-        cls.pb = ProbackupTest()
-        cls.backup_dir = os.path.join(cls.pb.tmp_path, module_name, 'backup')
-
-        cls.node = cls.pb.make_simple_node(
-            base_dir="{}/node".format(module_name),
+        self.node = self.pg_node.make_simple("node",
             set_replication=True,
-            initdb_params=['--data-checksums', '--auth-host=md5']
+            initdb_params=['--auth-host=md5'],
+            pg_options={'archive_timeout': '5s'},
         )
 
-        cls.username = cls.pb.get_username()
+        self.modify_pg_hba(self.node)
 
-        cls.modify_pg_hba(cls.node)
-
-        cls.pb.init_pb(cls.backup_dir)
-        cls.pb.add_instance(cls.backup_dir, cls.node.name, cls.node)
-        cls.pb.set_archiving(cls.backup_dir, cls.node.name, cls.node)
+        self.pb.init()
+        self.pb.add_instance(self.node.name, self.node)
+        self.pb.set_archiving(self.node.name, self.node)
         try:
-            cls.node.slow_start()
+            self.node.slow_start()
         except StartNodeException:
             raise unittest.skip("Node hasn't started")
 
-        if cls.pb.get_version(cls.node) < 100000:
-            cls.node.safe_psql(
-                "postgres",
-                "CREATE ROLE backup WITH LOGIN PASSWORD 'password'; "
-                "GRANT USAGE ON SCHEMA pg_catalog TO backup; "
-                "GRANT EXECUTE ON FUNCTION current_setting(text) TO backup; "
-                "GRANT EXECUTE ON FUNCTION pg_is_in_recovery() TO backup; "
-                "GRANT EXECUTE ON FUNCTION pg_start_backup(text, boolean, boolean) TO backup; "
-                "GRANT EXECUTE ON FUNCTION pg_stop_backup() TO backup; "
-                "GRANT EXECUTE ON FUNCTION pg_stop_backup(boolean) TO backup; "
-                "GRANT EXECUTE ON FUNCTION pg_create_restore_point(text) TO backup; "
-                "GRANT EXECUTE ON FUNCTION pg_switch_xlog() TO backup; "
-                "GRANT EXECUTE ON FUNCTION txid_current() TO backup; "
-                "GRANT EXECUTE ON FUNCTION txid_current_snapshot() TO backup; "
-                "GRANT EXECUTE ON FUNCTION txid_snapshot_xmax(txid_snapshot) TO backup;")
-        elif cls.pb.get_version(cls.node) < 150000:
-            cls.node.safe_psql(
+
+        version = self.pg_config_version
+        if version < 150000:
+            self.node.safe_psql(
                 "postgres",
                 "CREATE ROLE backup WITH LOGIN PASSWORD 'password'; "
                 "GRANT USAGE ON SCHEMA pg_catalog TO backup; "
@@ -233,7 +174,7 @@ class AuthTest(unittest.TestCase):
                 "GRANT EXECUTE ON FUNCTION txid_current_snapshot() TO backup; "
                 "GRANT EXECUTE ON FUNCTION txid_snapshot_xmax(txid_snapshot) TO backup;")
         else:
-            cls.node.safe_psql(
+            self.node.safe_psql(
                 "postgres",
                 "CREATE ROLE backup WITH LOGIN PASSWORD 'password'; "
                 "GRANT USAGE ON SCHEMA pg_catalog TO backup; "
@@ -247,37 +188,54 @@ class AuthTest(unittest.TestCase):
                 "GRANT EXECUTE ON FUNCTION txid_current_snapshot() TO backup; "
                 "GRANT EXECUTE ON FUNCTION txid_snapshot_xmax(txid_snapshot) TO backup;")
 
-        cls.pgpass_file = os.path.join(os.path.expanduser('~'), '.pgpass')
+        if version >= 150000:
+            home_dir = os.path.join(self.test_path, "home")
+            os.makedirs(home_dir, exist_ok=True)
+            self.test_env['HOME'] = home_dir
+            self.pgpass_file = os.path.join(home_dir, '.pgpass')
+            self.pgpass_file_lock = None
+        else:
+            # before PGv15 only true home dir were inspected.
+            # Since we can't have separate file per test, we have to serialize
+            # tests.
+            self.pgpass_file = os.path.join(os.path.expanduser('~'), '.pgpass')
+            self.pgpass_file_lock = self.pgpass_file + '~probackup_test_lock'
+            # have to lock pgpass by creating file in exclusive mode
+            for i in range(120):
+                try:
+                    open(self.pgpass_file_lock, "x").close()
+                except FileExistsError:
+                    time.sleep(1)
+                else:
+                    break
+            else:
+                raise TimeoutError("can't create ~/.pgpass~probackup_test_lock for 120 seconds")
 
-    # TODO move to object scope, replace module_name
-    @classmethod
-    def tearDownClass(cls):
-        cls.node.cleanup()
-        cls.pb.del_test_dir(module_name, '')
-
-    @unittest.skipIf(skip_test, "Module pexpect isn't installed. You need to install it.")
-    def setUp(self):
         self.pb_cmd = ['backup',
-                    '-B', self.backup_dir,
                     '--instance', self.node.name,
                     '-h', '127.0.0.1',
                     '-p', str(self.node.port),
                     '-U', 'backup',
                     '-d', 'postgres',
-                    '-b', 'FULL'
+                    '-b', 'FULL',
+                    '--no-sync'
                     ]
 
     def tearDown(self):
-        if "PGPASSWORD" in self.pb.test_env.keys():
-            del self.pb.test_env["PGPASSWORD"]
-
-        if "PGPASSWORD" in self.pb.test_env.keys():
-            del self.pb.test_env["PGPASSFILE"]
-
-        try:
-            os.remove(self.pgpass_file)
-        except OSError:
-            pass
+        super().tearDown()
+        if not self.pgpass_file_lock:
+            return
+        if hasattr(self, "pgpass_line") and os.path.exists(self.pgpass_file):
+            with open(self.pgpass_file, 'r') as fl:
+                lines = fl.readlines()
+            if self.pgpass_line in lines:
+                lines.remove(self.pgpass_line)
+            if len(lines) == 0:
+                os.remove(self.pgpass_file)
+            else:
+                with open(self.pgpass_file, 'w') as fl:
+                    fl.writelines(lines)
+        os.remove(self.pgpass_file_lock)
 
     def test_empty_password(self):
         """ Test case: PGPB_AUTH03 - zero password length """
@@ -313,13 +271,13 @@ class AuthTest(unittest.TestCase):
 
     def test_pgpassfile_env(self):
         """ Test case: PGPB_AUTH06 - set environment var PGPASSFILE """
-        path = os.path.join(self.pb.tmp_path, module_name, 'pgpass.conf')
+        path = os.path.join(self.test_path, 'pgpass.conf')
         line = ":".join(['127.0.0.1', str(self.node.port), 'postgres', 'backup', 'password'])
         self.create_pgpass(path, line)
-        self.pb.test_env["PGPASSFILE"] = path
+        self.test_env["PGPASSFILE"] = path
         self.assertEqual(
             "OK",
-            self.pb.show_pb(self.backup_dir, self.node.name, self.pb.run_pb(self.pb_cmd + ['-w']))["status"],
+            self.pb.show(self.node.name, self.pb.run(self.pb_cmd + ['-w']))["status"],
             "ERROR: Full backup status is not valid."
         )
 
@@ -329,16 +287,16 @@ class AuthTest(unittest.TestCase):
         self.create_pgpass(self.pgpass_file, line)
         self.assertEqual(
             "OK",
-            self.pb.show_pb(self.backup_dir, self.node.name, self.pb.run_pb(self.pb_cmd + ['-w']))["status"],
+            self.pb.show(self.node.name, self.pb.run(self.pb_cmd + ['-w']))["status"],
             "ERROR: Full backup status is not valid."
         )
 
     def test_pgpassword(self):
         """ Test case: PGPB_AUTH08 - set environment var PGPASSWORD """
-        self.pb.test_env["PGPASSWORD"] = "password"
+        self.test_env["PGPASSWORD"] = "password"
         self.assertEqual(
             "OK",
-            self.pb.show_pb(self.backup_dir, self.node.name, self.pb.run_pb(self.pb_cmd + ['-w']))["status"],
+            self.pb.show(self.node.name, self.pb.run(self.pb_cmd + ['-w']))["status"],
             "ERROR: Full backup status is not valid."
         )
 
@@ -346,16 +304,18 @@ class AuthTest(unittest.TestCase):
         """ Test case: PGPB_AUTH09 - Check priority between PGPASSWORD and .pgpass file"""
         line = ":".join(['127.0.0.1', str(self.node.port), 'postgres', 'backup', 'wrong_password'])
         self.create_pgpass(self.pgpass_file, line)
-        self.pb.test_env["PGPASSWORD"] = "password"
+        self.test_env["PGPASSWORD"] = "password"
         self.assertEqual(
             "OK",
-            self.pb.show_pb(self.backup_dir, self.node.name, self.pb.run_pb(self.pb_cmd + ['-w']))["status"],
+            self.pb.show(self.node.name, self.pb.run(self.pb_cmd + ['-w']))["status"],
             "ERROR: Full backup status is not valid."
         )
 
     def run_pb_with_auth(self, password=None, add_args = [], kill=False):
-        with spawn(self.pb.probackup_path, self.pb_cmd + add_args, encoding='utf-8', timeout=10) as probackup:
-            result = probackup.expect(u"Password for user .*:", 5)
+        cmd = [*self.pb_cmd, *add_args, *self.backup_dir.pb_args]
+        with spawn(self.probackup_path, cmd,
+                   encoding='utf-8', timeout=60, env=self.test_env) as probackup:
+            result = probackup.expect(u"Password for user .*:", 10)
             if kill:
                 probackup.kill(signal.SIGINT)
             elif result == 0:
@@ -366,8 +326,7 @@ class AuthTest(unittest.TestCase):
                 raise ExceptionPexpect("Other pexpect errors.")
 
 
-    @classmethod
-    def modify_pg_hba(cls, node):
+    def modify_pg_hba(self, node):
         """
         Description:
             Add trust authentication for user postgres. Need for add new role and set grant.
@@ -378,11 +337,12 @@ class AuthTest(unittest.TestCase):
         with open(hba_conf, 'r+') as fio:
             data = fio.read()
             fio.seek(0)
-            fio.write('host\tall\t%s\t127.0.0.1/0\ttrust\n%s' % (cls.username, data))
+            fio.write('host\tall\t%s\t127.0.0.1/0\ttrust\n%s' % (self.username, data))
 
 
     def create_pgpass(self, path, line):
-        with open(path, 'w') as passfile:
+        self.pgpass_line = line+"\n"
+        with open(path, 'a') as passfile:
             # host:port:db:username:password
-            passfile.write(line)
+            passfile.write(self.pgpass_line)
         os.chmod(path, 0o600)
